✨ Recommend instructor linking with warning on dashboard
インストラクター連携を推奨するスタンスに変更:
- 未連携時は黄色Alertで連携を促し、その場でSelect/Buttonから連携できるようにする
- 連携済み時は画面下部に「連携中・解除」を極薄グレーで最小表示し、解除操作を目立たせない
- 直近の勤務予定は動線として常に表示しつつ、未連携時のダミー行表示は無くす
